Soft-deprecated URLSession.enableAutomaticRegistration
